modify: 例外を共通化,クラス名を変更(IllegalParameterException)
