+++ src/matsu/num/statistics/kdeapp/exception/package-info.java
@@ -1,0 +1,11 @@
+/*
+ * Copyright © 2026 Matsuura Y.
+ * 
+ * This software is released under the MIT License.
+ * http://opensource.org/licenses/mit-license.php
+ */
+
+/**
+ * このアプリケーションにおける例外を扱うパッケージ.
+ */
+package matsu.num.statistics.kdeapp.exception;
