+++ src/matsu/num/statistics/kdeapp/kde1d/ResultDisplay.java
@@ -6,7 +6,7 @@
  */
 
 /*
- * 2026.2.15
+ * 2026.2.17
  */
 package matsu.num.statistics.kdeapp.kde1d;
 
@@ -14,7 +14,7 @@ import java.io.PrintStream;
 import java.io.PrintWriter;
 import java.util.Objects;
 
-import matsu.num.statistics.kdeapp.kde1d.exception.OutputException;
+import matsu.num.statistics.kdeapp.exception.OutputException;
 
 /**
  * 結果のディスプレイ出力を扱う.
