+++ src/matsu/num/statistics/kdeapp/kde1d/command/rule/CompositeRule.java
@@ -6,7 +6,7 @@
  */
 
 /*
- * 2026.2.11
+ * 2026.2.17
  */
 package matsu.num.statistics.kdeapp.kde1d.command.rule;
 
@@ -16,8 +16,8 @@ import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import matsu.num.statistics.kdeapp.exception.IllegalParameterException;
 import matsu.num.statistics.kdeapp.kde1d.command.ConsoleOptionCommand;
-import matsu.num.statistics.kdeapp.kde1d.exception.InvalidParameterException;
 
 /**
  * 複数のルールを結合した (バリデーションを "and" でつないだ) ルールを表現する.
@@ -42,7 +42,7 @@ final class CompositeRule implements CommandAssignmentRule {
     }
 
     /**
-     * @throws InvalidParameterException {@inheritDoc}
+     * @throws IllegalParameterException {@inheritDoc}
      * @throws NullPointerException {@inheritDoc}
      */
     @Override
