--- src/matsu/num/statistics/kdeapp/kde1d/exception/ApplicationException.java
@@ -1,56 +1,0 @@
-/*
- * Copyright © 2026 Matsuura Y.
- * 
- * This software is released under the MIT License.
- * http://opensource.org/licenses/mit-license.php
- */
-
-/*
- * 2026.2.10
- */
-package matsu.num.statistics.kdeapp.kde1d.exception;
-
-/**
- * アプリケーションに係る例外を扱うパッケージ.
- * 
- * @author Matsuura Y.
- */
-public class ApplicationException extends RuntimeException {
-
-    private static final long serialVersionUID = 1L;
-
-    /**
-     * 引数なしコンストラクタ.
-     */
-    public ApplicationException() {
-        super();
-    }
-
-    /**
-     * メッセージを与えるコンストラクタ.
-     * 
-     * @param message message
-     */
-    public ApplicationException(String message) {
-        super(message);
-    }
-
-    /**
-     * 原因を与えるコンストラクタ.
-     * 
-     * @param cause cause
-     */
-    public ApplicationException(Throwable cause) {
-        super(cause);
-    }
-
-    /**
-     * メッセージと原因を与えるコンストラクタ.
-     * 
-     * @param message message
-     * @param cause cause
-     */
-    public ApplicationException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
