--- src/matsu/num/statistics/kdeapp/kde1d/exception/package-info.java
@@ -1,15 +1,0 @@
-/*
- * Copyright © 2026 Matsuura Y.
- * 
- * This software is released under the MIT License.
- * http://opensource.org/licenses/mit-license.php
- */
-
-/**
- * kde1d の例外を扱うパッケージ.
- * 
- * <p>
- * 有用性を十分に吟味し, プロジェクト共通の例外パッケージに昇格させる.
- * </p>
- */
-package matsu.num.statistics.kdeapp.kde1d.exception;
