+++ src/matsu/num/statistics/kdeapp/kde2d/ResultOutput.java
@@ -21,7 +21,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Objects;
 
-import matsu.num.statistics.kdeapp.kde2d.exception.OutputException;
+import matsu.num.statistics.kdeapp.exception.OutputException;
 
 /**
  * 結果の外部出力を扱う.
