+++ test/matsu/num/statistics/kdeapp/kde1d/GaussianStandardKde1dCalculatorTest.java
@@ -11,7 +11,7 @@ import org.junit.Test;
 import org.junit.experimental.runners.Enclosed;
 import org.junit.runner.RunWith;
 
-import matsu.num.statistics.kdeapp.kde1d.exception.CalculationException;
+import matsu.num.statistics.kdeapp.exception.CalculationException;
 
 /**
  * {@link GaussianStandardKde1dCalculator} のテスト.
