+++ test/matsu/num/statistics/kdeapp/kde2d/GaussianStandardKde2dCalculatorTest.java
@@ -11,7 +11,7 @@ import org.junit.Test;
 import org.junit.experimental.runners.Enclosed;
 import org.junit.runner.RunWith;
 
-import matsu.num.statistics.kdeapp.kde2d.exception.CalculationException;
+import matsu.num.statistics.kdeapp.exception.CalculationException;
 
 /**
  * {@link GaussianStandardKde2dCalculator} のテスト.
